backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.1
Markdown==3.3.4
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.6.0
